Interconnected the js files with import/export
